requirements.txt
python-telegram-bot>=20.0
pillow>=10.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0